Island: Move IslandModeEnum to monkey_island.cc.models
