+++ monkey/monkey_island/cc/models/__init__.py
@@ -8,4 +8,4 @@ from .monkey_ttl import MonkeyTtl
 from .pba_results import PbaResults
 from monkey_island.cc.models.report.report import Report
 from .stolen_credentials import StolenCredentials
-from .simulation import Simulation, SimulationSchema
+from .simulation import Simulation, SimulationSchema, IslandModeEnum
