+++ monkey/monkey_island/cc/resources/island_mode.py
@@ -3,10 +3,10 @@ import logging
 
 from flask import make_response, request
 
+from monkey_island.cc.models import IslandModeEnum
 from monkey_island.cc.resources.AbstractResource import AbstractResource
 from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services import IslandModeService
-from monkey_island.cc.services.mode.mode_enum import IslandModeEnum
 
 logger = logging.getLogger(__name__)
